==> 1.InitialScraping.ipynb <==
import GetOldTweets3 as got
import pandas as pd 
import time

def scrape_all_tweets_backwards(start, search_term):
    counter = 129482998875612  # I went back 10k, should work now, hopefully capture the first one
    since = start
    for i in range(1000):
        tweetCriteria = got.manager.TweetCriteria().setQuerySearch(f'{search_term} max_id:{counter}')\
                                                        .setSince(start)\
                                                        .setMaxTweets(1000)
        new_scrapes = got.manager.TweetManager.getTweets(tweetCriteria)
        df = pd.DataFrame(t.__dict__ for t in new_scrapes)
        counter = df.tail(1)['id'].values[0]   # HALEYYYYYY FTW
        date = df.tail(1)['date'].values[0]
        print(counter)
        print(date)
        print(f'Just scraped 1000 tweets and waiting to go again.')
        df.to_csv('./data/corpus2.csv', index=False, mode='a')
        time.sleep(60)
        

# scrape_all_tweets_backwards("2019-07-02", "#rip45")

scrape_all_tweets_backwards("2020-08-10", "#wrongtrump")
