==> 2.EDA.ipynb <==
import pandas as pd

wt = pd.read_csv('data/#wrongtrump.csv')

wt.shape

wt.tail()

wt.info()

wt['retweets'].tail()

columns_to_drop = ['to', 'retweets', 'favorites', 'replies', 'id', 'permalink', 'author_id', 'formatted_date', 'hashtags', 'mentions', 'geo', 'urls']

wt = wt.drop(labels=columns_to_drop, axis=1)

### Zero Null Values
### Need to drop Duplicates

CitizenCee = wt[wt['username'] == 'CitizenCee']

CitizenCee.head()

wt = wt[wt['username'] != 'username']

wt.shape

meow10107 = wt[wt['username'] == 'meow10107']

meow10107.head(50)

### This one is very interesting and extremely botty and early

wt['username'].value_counts().head(15)

AnkerOfficial = wt[wt['username'] == 'AnkerOfficial']
AnkerOfficial.head(50)

### Clearly an Ad, can clear these

JoinDeepRock = wt[wt['username'] == 'JoinDeepRock']
JoinDeepRock.head(50)

### This is also an AD

BHG = wt[wt['username'] == 'BHG']
BHG.head(50)

#also an ad

meToyou99961431 = wt[wt['username'] == 'meToyou99961431']
meToyou99961431.head(50)


TalbertSwan = wt[wt['username'] == 'TalbertSwan']
TalbertSwan.head(50)

meToyou99961431 = wt[wt['username'] == 'meToyou99961431']
meToyou99961431.head(50)
