==> 3.Training Scrape.ipynb <==
### Imports

import pandas as pd
import snscrape.modules.twitter
import csv
import os

### Function to scrape Training Users
##### 1. Open new CSV file, name it after the User, place it in appropriate folder.
##### 2. Scrape all tweets from User and write them to their own row.

def get_training_user(handle, color):
    header = ['text', 'user', 'date']
    tweets=snscrape.modules.twitter.TwitterUserScraper(username=handle)
    filename = f'{color}/{handle}.csv'
    with open(filename, 'a', encoding='utf-8') as f:
            wr=csv.writer(f, lineterminator = '\n')
            wr.writerow(i for i in header)
            for i in tweets.get_items():
                wr.writerow([i.content, i.user, i.date])
    print(f'Great, all the tweets from {handle}')

def add_to_red_list(name):
    if f'{name}.csv' in os.listdir('red2'):
        print("already got it")
    elif f'red_{name}.csv' in os.listdir('red'):
        print("we got it in the first batch of reds")
    else:
        return get_training_user(name, 'red2')

    
def add_to_blue_list(name):
    if f'{name}.csv' in os.listdir('blue2'):
        print("already got it")
    elif f'blue_{name}.csv' in os.listdir('blue'):
        print("we got it in the first batch of blues")
    else:
        return get_training_user(name, 'blue2')

add_to_red_list('huckabeeontbn')

add_to_blue_list('aoc')
