==> 4.Update Training Docs.ipynb <==
### Imports

import os
import pandas as pd
from tqdm import tqdm

### Function that takes a folder of scraped users and their label, and adds them to the training corpus

def add_to_training_docs(folder, color):
    new_corpus = pd.DataFrame(columns=['user', 'doc', 'label'])
    for file in tqdm(os.listdir(folder)[1:]):
        df = pd.read_csv(f"{folder}/{file}", header=None, engine='python')
        stringaling = ''
        for tweet in df[0]:
            stringaling = stringaling + ' ' + tweet
        new_corpus['user'] = [file.replace('.csv', '')]
        new_corpus['doc'] = [stringaling]
        new_corpus['label'] = [color]
        training_documents = pd.read_csv('training_documents.csv', index_col=0)
        new_df = pd.concat([training_documents, new_corpus])
        new_df.to_csv('training_documents.csv')      
        print(f'{file} has been added to the new corpus')
    

add_to_training_docs('red2', 'red')

add_to_training_docs('blue2', 'blue')
