==> 6.Model Selection.ipynb <==
### Imports and Bringing in Custom Functions

import pandas as pd 
import numpy as np 
from tqdm import tqdm
from sklearn.svm import SVC
from sklearn.naive_bayes import GaussianNB
from sklearn.neighbors import KNeighborsClassifier
from gensim.models.doc2vec import Doc2Vec, TaggedDocument
from gensim.utils import simple_preprocess
from sklearn.model_selection import GridSearchCV, train_test_split
from sklearn.metrics import accuracy_score
from sklearn.linear_model import LogisticRegression
tqdm.pandas(desc="Applying")

def train_doc2vec_model(dayfray, name_of_model):
    processed = dayfray.apply(lambda x: TaggedDocument(words=(x['doc']), tags=[x['label']]), axis=1)
    print('processed')
    sents = processed.values
    new_model = Doc2Vec(sents, vector_size=300, epochs=40, dm=0, min_count=300)
    print('trained')
    new_model.save(name_of_model)
    print(f'saved {name_of_model}')

##### 1. Bring in training corpus
##### 2. Create X and y
##### 3. Train, Test split
##### 4. Create dataframes that I can feed to Doc2Vec

df = pd.read_csv("training_documents.csv", index_col=0)
df.head()
type(df['doc'][1])

df.head()

df['doc'] = df['doc'].progress_apply(lambda x: simple_preprocess(x))
type(df['doc'][1])

df['label'].value_counts()

df['label'] = df['label'].map({"red": 0, "blue": 1})
X = df['doc']
y = df['label']
print(y.value_counts(normalize=True))
X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y, test_size=0.2, random_state=32)
df_train = pd.DataFrame()
df_test = pd.DataFrame()
df_train['doc'] = X_train
df_test['doc'] = X_test
df_train['label'] = y_train
df_test['label'] = y_test

### Train Doc2Vec on training set

train_doc2vec_model(df_train, 'test_doc2vec')


### Infer Vectors for training set and testing set 

model = Doc2Vec.load('test_doc2vec')

df_train['vector'] = [model.infer_vector(list(x)) for x in tqdm(df_train['doc'])]

df_test['vector'] = [model.infer_vector(list(x)) for x in tqdm(df_test['doc'])]

df_test.shape

df_train.shape

## Begin Testing Models

### Support Vector Classifier

svc_test = SVC(probability=True)
svc_test.fit(list(df_train['vector']), df_train['label'])

train_hat = svc_test.predict(list(df_train['vector']))
print(accuracy_score(df_train['label'], train_hat))
test_hat = svc_test.predict(list(df_test['vector']))
print(accuracy_score(df_test['label'], test_hat))

df_test.head()

### Logistic Regression

logreg_test = LogisticRegression()
logreg_test.fit(list(df_train['vector']), df_train['label'])
train_hat = logreg_test.predict(list(df_train['vector']))
print(accuracy_score(df_train['label'], train_hat))
test_hat = logreg_test.predict(list(df_test['vector']))
print(accuracy_score(df_test['label'], test_hat))

### K Nearest Neighbor

knn_test = KNeighborsClassifier(n_neighbors=5)
knn_test.fit(list(df_train['vector']), df_train['label'])
train_hat = knn_test.predict(list(df_train['vector']))
print(accuracy_score(df_train['label'], train_hat))
test_hat = knn_test.predict(list(df_test['vector']))
print(accuracy_score(df_test['label'], test_hat))

### Quick Attempt at Clustering

from sklearn.cluster import KMeans
kmeans = KMeans(n_clusters=2)
cluster_preds = kmeans.fit_predict(list(df_train['vector']))

print(accuracy_score(df_train['label'], cluster_preds))

### Train Final Doc2Vec Model and Final SVC Model

train_doc2vec_model(df, 'final_doc2vec')

final_doc2vec = Doc2Vec.load('final_doc2vec')

df['vector'] = [final_doc2vec.infer_vector(list(x)) for x in tqdm(df['doc'])]

svc_final = SVC(probability=True)
svc_final.fit(list(df['vector']), df['label'])

import pickle


svc_filename = "svc_final_model.pkl"


with open(svc_filename, 'wb') as file:
    pickle.dump(svc_final, file)
